+++ admin/modules/blog/post-new.php
@@ -16,19 +16,25 @@ if( isset($_POST['postSubmit']) ) {
     $post->content = $_POST['content'];
     $post->timestamp= time();
 
-    // Если передано изображение - уменьшаем, сохраняем в папку
-    $coverFileName = saveUploadedImg('cover', [600, 300], 12, 'blog', [1110, 460], [290, 230]);
-     // Если новое изображение успешно загружено 
-    if ($coverFileName) {
-      // Удаляем старое изображение
-      unlink(ROOT . 'usercontent/blog/' . $post->cover);
-      unlink(ROOT . 'usercontent/blog/' . $post->coverSmall);
+    // Если передано изображение - уменьшаем, сохраняем, записываем в БД
+    if ( isset($_FILES['cover']['name']) && $_FILES['cover']['tmp_name'] !== '') {
+      //Если передано изображение - уменьшаем, сохраняем файлы в папку
+      $coverFileName = saveUploadedImg('cover', [600, 300], 12, 'blog', [1110, 460], [290, 230]);
+
+      // Если новое изображение успешно загружено 
+      if ($coverFileName) {
+        // Удаляем старое изображение
+        unlink(ROOT . 'usercontent/blog/' . $post->cover);
+        unlink(ROOT . 'usercontent/blog/' . $post->coverSmall);
+      }
+
+      // Записываем имя файлов в БД
+      $post->cover = $coverFileName[0];
+      $post->coverSmall = $coverFileName[1];
     }
-    // Записываем имя файлов в БД
-    $post->cover = $coverFileName[0];
-    $post->coverSmall = $coverFileName[1];
 
     R::store($post);
+
     $_SESSION['success'][] = ['title' => 'Пост успешно добавлен'];
     header('Location: ' . HOST . 'admin/blog');
     exit();
